requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
slowapi==0.1.9
python-multipart==0.0.6
pydantic[email]==2.5.0
email-validator==2.1.0 